Simple ARD Statistics

Compute Analysis Results Data (ARD) for simple summary statistics from continuous and categorical data.

ard_continuous(
  data,
  by = dplyr::group_vars(data),
  statistics = NULL,
  include = everything()
)

ard_categorical(data, by = dplyr::group_vars(data), include = everything())

Arguments

data

a data frame

by

columns to compute statistics by. Default are the columns returned by dplyr::group_vars(data).

statistics

a named list of functions that return a summary statistic, e.g. list(mpg = list(mean = \(x) mean(x, na.rm = TRUE)))

include

columns to include in summaries. Default is everything().

Value

a data frame

Examples

ard_continuous(mtcars, by = cyl, include = c(mpg, hp))
#> # A tibble: 42 × 5
#>    strata1 strata1_levels variable stat_name statistic
#>    <chr>            <dbl> <chr>    <chr>     <list>   
#>  1 cyl                  6 mpg      N         <int [1]>
#>  2 cyl                  6 mpg      N_miss    <int [1]>
#>  3 cyl                  6 mpg      N_tot     <int [1]>
#>  4 cyl                  6 mpg      mean      <dbl [1]>
#>  5 cyl                  6 mpg      sd        <dbl [1]>
#>  6 cyl                  6 mpg      min       <dbl [1]>
#>  7 cyl                  6 mpg      max       <dbl [1]>
#>  8 cyl                  6 hp       N         <int [1]>
#>  9 cyl                  6 hp       N_miss    <int [1]>
#> 10 cyl                  6 hp       N_tot     <int [1]>
#> # ℹ 32 more rows
ard_categorical(mtcars, by = cyl, include = c(am, gear))
#> # A tibble: 46 × 6
#>    strata1 strata1_levels variable variable_level stat_name statistic
#>    <chr>            <dbl> <chr>    <list>         <chr>     <list>   
#>  1 cyl                  6 am       <dbl [1]>      n         <int [1]>
#>  2 cyl                  6 am       <dbl [1]>      p         <dbl [1]>
#>  3 cyl                  6 am       <dbl [1]>      n         <int [1]>
#>  4 cyl                  6 am       <dbl [1]>      p         <dbl [1]>
#>  5 cyl                  4 am       <dbl [1]>      n         <int [1]>
#>  6 cyl                  4 am       <dbl [1]>      p         <dbl [1]>
#>  7 cyl                  4 am       <dbl [1]>      n         <int [1]>
#>  8 cyl                  4 am       <dbl [1]>      p         <dbl [1]>
#>  9 cyl                  8 am       <dbl [1]>      n         <int [1]>
#> 10 cyl                  8 am       <dbl [1]>      p         <dbl [1]>
#> # ℹ 36 more rows