Simple ARD Statistics

Compute Analysis Results Data (ARD) for simple summary statistics from continuous and categorical data.

Usage

ard_continuous(
  data,
  by = dplyr::group_vars(data),
  variables = everything(),
  statistics = NULL
)

ard_categorical(data, by = dplyr::group_vars(data), variables = everything())

Arguments

data

a data frame

by

columns to compute statistics by. Default are the columns returned by dplyr::group_vars(data).

variables

columns to include in summaries. Default is everything().

statistics

a named list of functions that return a summary statistic, e.g. list(mpg = list(mean = \(x) mean(x, na.rm = TRUE)))

Value

a data frame

Examples

ard_continuous(mtcars, by = cyl, variables = c("mpg", "hp"))
#> # A tibble: 42 × 8
#>    strata1 strata1_level variable stat_name statistic warning error  context   
#>  * <chr>   <list>        <chr>    <chr>     <list>    <list>  <list> <chr>     
#>  1 cyl     <dbl [1]>     mpg      N         <int [1]> <NULL>  <NULL> continuous
#>  2 cyl     <dbl [1]>     mpg      N_miss    <int [1]> <NULL>  <NULL> continuous
#>  3 cyl     <dbl [1]>     mpg      N_tot     <int [1]> <NULL>  <NULL> continuous
#>  4 cyl     <dbl [1]>     mpg      mean      <dbl [1]> <NULL>  <NULL> continuous
#>  5 cyl     <dbl [1]>     mpg      sd        <dbl [1]> <NULL>  <NULL> continuous
#>  6 cyl     <dbl [1]>     mpg      min       <dbl [1]> <NULL>  <NULL> continuous
#>  7 cyl     <dbl [1]>     mpg      max       <dbl [1]> <NULL>  <NULL> continuous
#>  8 cyl     <dbl [1]>     hp       N         <int [1]> <NULL>  <NULL> continuous
#>  9 cyl     <dbl [1]>     hp       N_miss    <int [1]> <NULL>  <NULL> continuous
#> 10 cyl     <dbl [1]>     hp       N_tot     <int [1]> <NULL>  <NULL> continuous
#> # ℹ 32 more rows
ard_categorical(mtcars, by = cyl, variables = c("am", "gear"))
#> # A tibble: 48 × 9
#>    strata1 strata1_level variable variable_level warning error  context    
#>  * <chr>   <list>        <chr>    <list>         <list>  <list> <chr>      
#>  1 cyl     <dbl [1]>     am       <dbl [1]>      <NULL>  <NULL> categorical
#>  2 cyl     <dbl [1]>     am       <dbl [1]>      <NULL>  <NULL> categorical
#>  3 cyl     <dbl [1]>     am       <dbl [1]>      <NULL>  <NULL> categorical
#>  4 cyl     <dbl [1]>     am       <dbl [1]>      <NULL>  <NULL> categorical
#>  5 cyl     <dbl [1]>     am       <dbl [1]>      <NULL>  <NULL> categorical
#>  6 cyl     <dbl [1]>     am       <dbl [1]>      <NULL>  <NULL> categorical
#>  7 cyl     <dbl [1]>     am       <dbl [1]>      <NULL>  <NULL> categorical
#>  8 cyl     <dbl [1]>     am       <dbl [1]>      <NULL>  <NULL> categorical
#>  9 cyl     <dbl [1]>     am       <dbl [1]>      <NULL>  <NULL> categorical
#> 10 cyl     <dbl [1]>     am       <dbl [1]>      <NULL>  <NULL> categorical
#> # ℹ 38 more rows
#> # ℹ 2 more variables: stat_name <chr>, statistic <list>